Move trace recorder test classes from `impl` to `runner` package
