+++ integration-test/trace-recorder/src/main/org/jetbrains/trace/recorder/test/runner/GenerateTests.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.trace.recorder.test.impl
+package org.jetbrains.trace.recorder.test.runner
 
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
